--- Packages/UGF.WebRequests/Runtime/WebRequestCookieUtility.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 62781e6fbcc1467fbca00c561e6e9089
-timeCreated: 1660408081
